refactor: move tests related to variables
