+++ zeebe/engine/src/test/java/io/camunda/zeebe/engine/processing/bpmn/activity/listeners/task/TaskListenerTest.java
@@ -43,7 +43,6 @@ import java.util.function.Predicate;
 import java.util.function.UnaryOperator;
 import org.junit.Before;
 import org.junit.ClassRule;
-import org.junit.Ignore;
 import org.junit.Rule;
 import org.junit.Test;
 
@@ -51,7 +50,6 @@ public class TaskListenerTest {
 
   @ClassRule public static final EngineRule ENGINE = EngineRule.singlePartition();
 
-  private static final String USER_TASK_ELEMENT_ID = "my_user_task";
   private static final List<String> ALL_CORRECTABLE_ATTRIBUTES =
       List.of(
           UserTaskRecord.ASSIGNEE,
@@ -71,102 +69,6 @@ public class TaskListenerTest {
   @Before
   public void setup() {
     listenerType = "my_listener_" + UUID.randomUUID();
-  }
-
-  @Test
-  @Ignore(
-      "Ignored due to task listener job completion rejection when variables payload is provided (issue #24056). Re-enable after implementing issue #23702.")
-  public void shouldMakeVariablesFromPreviousTaskListenersAvailableToSubsequentListeners() {
-    final long processInstanceKey =
-        helper.createProcessInstance(
-            helper.createProcessWithCompletingTaskListeners(listenerType, listenerType + "_2"));
-
-    ENGINE.userTask().ofInstance(processInstanceKey).complete();
-
-    // when
-    ENGINE
-        .job()
-        .ofInstance(processInstanceKey)
-        .withType(listenerType)
-        .withVariable("listener_1_var", "foo")
-        .complete();
-
-    // then: `listener_1_var` variable accessible in subsequent TL
-    final var jobActivated = helper.activateJob(processInstanceKey, listenerType + "_2");
-    assertThat(jobActivated.getVariables()).contains(entry("listener_1_var", "foo"));
-  }
-
-  @Test
-  @Ignore(
-      "Ignored due to task listener job completion rejection when variables payload is provided (issue #24056). Re-enable after implementing issue #23702.")
-  public void shouldNotExposeTaskListenerVariablesOutsideUserTaskScope() {
-    // given: deploy a process with a user task having complete TL and service task following it
-    final long processInstanceKey =
-        helper.createProcessInstance(
-            helper.createProcess(
-                p ->
-                    p.userTask(
-                            USER_TASK_ELEMENT_ID,
-                            t ->
-                                t.zeebeUserTask()
-                                    .zeebeAssignee("foo")
-                                    .zeebeTaskListener(l -> l.completing().type(listenerType)))
-                        .serviceTask(
-                            "subsequent_service_task",
-                            tb -> tb.zeebeJobType("subsequent_service_task"))));
-
-    ENGINE.userTask().ofInstance(processInstanceKey).complete();
-
-    // when: complete TL job with a variable 'my_listener_var'
-    ENGINE
-        .job()
-        .ofInstance(processInstanceKey)
-        .withType(listenerType)
-        .withVariable("my_listener_var", "bar")
-        .complete();
-
-    // then: assert the variable 'my_listener_var' isn't accessible in the subsequent element
-    final var subsequentServiceTaskJob =
-        helper.activateJob(processInstanceKey, "subsequent_service_task");
-    assertThat(subsequentServiceTaskJob.getVariables()).doesNotContainKey("my_listener_var");
-    helper.completeJobs(processInstanceKey, "subsequent_service_task");
-  }
-
-  @Test
-  @Ignore(
-      "Ignored due to task listener job completion rejection when variables payload is provided (issue #24056). Re-enable after implementing issue #23702.")
-  public void shouldAllowTaskListenerVariablesInUserTaskOutputMappings() {
-    // given: deploy a process with a user task having complete TL and service task following it
-    final long processInstanceKey =
-        helper.createProcessInstance(
-            helper.createProcess(
-                p ->
-                    p.userTask(
-                            USER_TASK_ELEMENT_ID,
-                            t ->
-                                t.zeebeUserTask()
-                                    .zeebeAssignee("foo")
-                                    .zeebeTaskListener(l -> l.completing().type(listenerType))
-                                    .zeebeOutput("=my_listener_var+\"_abc\"", "userTaskOutput"))
-                        .serviceTask(
-                            "subsequent_service_task",
-                            tb -> tb.zeebeJobType("subsequent_service_task"))));
-
-    ENGINE.userTask().ofInstance(processInstanceKey).complete();
-
-    // when: complete TL job with a variable 'my_listener_var'
-    ENGINE
-        .job()
-        .ofInstance(processInstanceKey)
-        .withType(listenerType)
-        .withVariable("my_listener_var", "bar")
-        .complete();
-
-    // then: assert the variable 'userTaskOutput' is accessible in the subsequent element
-    final var subsequentServiceTaskJob =
-        helper.activateJob(processInstanceKey, "subsequent_service_task");
-    assertThat(subsequentServiceTaskJob.getVariables()).containsEntry("userTaskOutput", "bar_abc");
-    helper.completeJobs(processInstanceKey, "subsequent_service_task");
   }
 
   @Test
