+++ zeebe/engine/src/test/java/io/camunda/zeebe/engine/processing/bpmn/activity/listeners/task/TaskListenerVariablesTest.java
@@ -7,17 +7,24 @@
  */
 package io.camunda.zeebe.engine.processing.bpmn.activity.listeners.task;
 
+import static java.util.Map.entry;
+import static org.assertj.core.api.Assertions.assertThat;
+
 import io.camunda.zeebe.engine.util.EngineRule;
 import io.camunda.zeebe.test.util.record.RecordingExporterTestWatcher;
 import java.util.UUID;
 import org.junit.Before;
 import org.junit.ClassRule;
+import org.junit.Ignore;
 import org.junit.Rule;
+import org.junit.Test;
 import org.junit.rules.TestWatcher;
 
 /** Tests verifying that a task listener can access and provide variables. */
 public class TaskListenerVariablesTest {
   @ClassRule public static final EngineRule ENGINE = EngineRule.singlePartition();
+  private static final String USER_TASK_ELEMENT_ID = "my_user_task";
+
   @Rule public final TestWatcher watcher = new RecordingExporterTestWatcher();
   private final TaskListenerTestHelper helper = new TaskListenerTestHelper(ENGINE);
 
@@ -26,5 +33,101 @@ public class TaskListenerVariablesTest {
   @Before
   public void setup() {
     listenerType = "my_listener_" + UUID.randomUUID();
+  }
+
+  @Test
+  @Ignore(
+      "Ignored due to task listener job completion rejection when variables payload is provided (issue #24056). Re-enable after implementing issue #23702.")
+  public void shouldMakeVariablesFromPreviousTaskListenersAvailableToSubsequentListeners() {
+    final long processInstanceKey =
+        helper.createProcessInstance(
+            helper.createProcessWithCompletingTaskListeners(listenerType, listenerType + "_2"));
+
+    ENGINE.userTask().ofInstance(processInstanceKey).complete();
+
+    // when
+    ENGINE
+        .job()
+        .ofInstance(processInstanceKey)
+        .withType(listenerType)
+        .withVariable("listener_1_var", "foo")
+        .complete();
+
+    // then: `listener_1_var` variable accessible in subsequent TL
+    final var jobActivated = helper.activateJob(processInstanceKey, listenerType + "_2");
+    assertThat(jobActivated.getVariables()).contains(entry("listener_1_var", "foo"));
+  }
+
+  @Test
+  @Ignore(
+      "Ignored due to task listener job completion rejection when variables payload is provided (issue #24056). Re-enable after implementing issue #23702.")
+  public void shouldNotExposeTaskListenerVariablesOutsideUserTaskScope() {
+    // given: deploy a process with a user task having complete TL and service task following it
+    final long processInstanceKey =
+        helper.createProcessInstance(
+            helper.createProcess(
+                p ->
+                    p.userTask(
+                            USER_TASK_ELEMENT_ID,
+                            t ->
+                                t.zeebeUserTask()
+                                    .zeebeAssignee("foo")
+                                    .zeebeTaskListener(l -> l.completing().type(listenerType)))
+                        .serviceTask(
+                            "subsequent_service_task",
+                            tb -> tb.zeebeJobType("subsequent_service_task"))));
+
+    ENGINE.userTask().ofInstance(processInstanceKey).complete();
+
+    // when: complete TL job with a variable 'my_listener_var'
+    ENGINE
+        .job()
+        .ofInstance(processInstanceKey)
+        .withType(listenerType)
+        .withVariable("my_listener_var", "bar")
+        .complete();
+
+    // then: assert the variable 'my_listener_var' isn't accessible in the subsequent element
+    final var subsequentServiceTaskJob =
+        helper.activateJob(processInstanceKey, "subsequent_service_task");
+    assertThat(subsequentServiceTaskJob.getVariables()).doesNotContainKey("my_listener_var");
+    helper.completeJobs(processInstanceKey, "subsequent_service_task");
+  }
+
+  @Test
+  @Ignore(
+      "Ignored due to task listener job completion rejection when variables payload is provided (issue #24056). Re-enable after implementing issue #23702.")
+  public void shouldAllowTaskListenerVariablesInUserTaskOutputMappings() {
+    // given: deploy a process with a user task having complete TL and service task following it
+    final long processInstanceKey =
+        helper.createProcessInstance(
+            helper.createProcess(
+                p ->
+                    p.userTask(
+                            USER_TASK_ELEMENT_ID,
+                            t ->
+                                t.zeebeUserTask()
+                                    .zeebeAssignee("foo")
+                                    .zeebeTaskListener(l -> l.completing().type(listenerType))
+                                    .zeebeOutput("=my_listener_var+\"_abc\"", "userTaskOutput"))
+                        .serviceTask(
+                            "subsequent_service_task",
+                            tb -> tb.zeebeJobType("subsequent_service_task"))));
+
+    ENGINE.userTask().ofInstance(processInstanceKey).complete();
+
+    // when: complete TL job with a variable 'my_listener_var'
+    ENGINE
+        .job()
+        .ofInstance(processInstanceKey)
+        .withType(listenerType)
+        .withVariable("my_listener_var", "bar")
+        .complete();
+
+    // then: assert the variable 'userTaskOutput' is accessible in the subsequent element
+    final var subsequentServiceTaskJob =
+        helper.activateJob(processInstanceKey, "subsequent_service_task");
+    assertThat(subsequentServiceTaskJob.getVariables()).containsEntry("userTaskOutput", "bar_abc");
+    helper.completeJobs(processInstanceKey, "subsequent_service_task");
   }
 }
